\name{artfima-package}
\alias{artfima-package}

\docType{package}
\title{\packageTitle{artfima}}
\description{\packageDescription{artfima}}

\details{
\packageDESCRIPTION{artfima}
\packageIndices{artfima}
}


\author{
\packageAuthor{artfima}
Maintainer: \packageMaintainer{artfima}
}

\references{TBA}

\keyword{ package }
\keyword{ts}

\seealso{
\code{\link{ltsa}}
}

\examples{
artfima(rnorm(100))
}
