\name{Huron}
\alias{Huron}
\docType{data}
\title{
Mean level in July of Lake Huron from 1860 to 1986
}
\description{
Mean July average water surface elevation, in feet, IGLD (1955) 
for Harbor Beach, Michigan, on Lake Huron, Station 5014.  1860--1986.
}
\usage{data("Huron")}
\format{
  The format is:
 Time-Series [1:127] from 1860 to 1986: 582 582 581 581 580 ...
}
\source{
Great Lakes Water Levels, 1860-1986. U.S. Dept. of Commerce,
National Oceanic and Atmospheric Administration, National Ocean Survey.
}
\examples{
data(Huron)
plot(Huron)
}
\keyword{datasets}