\name{Ogden}
\alias{Ogden}
\docType{data}
\title{
ST. Lawrence River at Ogdensburg, N.Y., 1860-1957, yearly flow 
}
\description{
flow is in cms
}
\usage{data("Ogden")}
\format{
  The format is:
 Time-Series [1:97] from 1860 to 1956: 7788 8040 7733 7528 7528 ...
 - attr(*, "title")= chr "St. Lawrence river at Ogdensburg, N.Y., 1860-1957, yearly flow"
}

\source{
"Time Series Modelling of Water Resources and Environmental Systems" by Keith W. Hipel and A. Ian McLeod.
}
\references{
"Time Series Modelling of Water Resources and Environmental Systems" by Keith W. Hipel and A. Ian McLeod.
}

\examples{
data(Ogden)
plot(Ogden)
}
\keyword{datasets}
